# gateway/docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    email: support@erp-system.com
    name: ERP System Support
  description: Enterprise Resource Planning System API Gateway. Provides unified REST
    API access to Inventory, Order, Finance, and HR microservices.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ERP System API
  version: "1.0"
paths:
...
  /api/v1/hr/attendance:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List attendance records
      tags:
      - HR - Attendance
  /api/v1/hr/attendance/clock-in:
    post:
      consumes:
      - application/json
      parameters:
      - description: Clock-in data
        in: body
        name: data
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Clock in attendance
      tags:
      - HR - Attendance
  /api/v1/hr/attendance/clock-out:
    post:
      consumes:
      - application/json
      parameters:
      - description: Clock-out data
        in: body
        name: data
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Clock out attendance
      tags:
      - HR - Attendance
  /api/v1/hr/dashboard:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get HR dashboard
      tags:
      - HR - Dashboard
  /api/v1/hr/departments:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all departments
      tags:
      - HR - Departments
    post:
      consumes:
      - application/json
      parameters:
      - description: Department data
        in: body
        name: dept
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new department
      tags:
      - HR - Departments
  /api/v1/hr/departments/{id}:
    get:
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a department by ID
      tags:
      - HR - Departments
  /api/v1/hr/employees:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all employees
      tags:
      - HR - Employees
    post:
      consumes:
      - application/json
      parameters:
      - description: Employee data
        in: body
        name: employee
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new employee
      tags:
      - HR - Employees
  /api/v1/hr/employees/{id}:
    get:
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get an employee by ID
      tags:
      - HR - Employees
    put:
      consumes:
      - application/json
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated employee data
        in: body
        name: employee
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update an employee
      tags:
      - HR - Employees
  /api/v1/hr/employees/{id}/terminate:
    post:
      consumes:
      - application/json
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      - description: Termination data
        in: body
        name: data
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Terminate an employee
      tags:
      - HR - Employees
  /api/v1/hr/leave/{id}/approve:
    post:
      parameters:
      - description: Leave Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Approve a leave request
      tags:
      - HR - Leave Management
  /api/v1/hr/leave/{id}/reject:
    post:
      parameters:
      - description: Leave Request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Reject a leave request
      tags:
      - HR - Leave Management
  /api/v1/hr/leave/balance/{employeeId}:
    get:
      parameters:
      - description: Employee ID
        in: path
        name: employeeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get leave balance for an employee
      tags:
      - HR - Leave Management
  /api/v1/hr/leave/request:
    post:
      consumes:
      - application/json
      parameters:
      - description: Leave request data
        in: body
        name: leave
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Submit a leave request
      tags:
      - HR - Leave Management
  /api/v1/hr/leave/requests:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all leave requests
      tags:
      - HR - Leave Management
  /api/v1/hr/payroll:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all payroll runs
      tags:
      - HR - Payroll
    post:
      consumes:
      - application/json
      parameters:
      - description: Payroll run data
        in: body
        name: payroll
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a payroll run
      tags:
      - HR - Payroll
  /api/v1/hr/payroll/{runId}:
    get:
      parameters:
      - description: Payroll Run ID
        in: path
        name: runId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a payroll run by ID
      tags:
      - HR - Payroll
  /api/v1/hr/payroll/{runId}/approve:
    post:
      parameters:
      - description: Payroll Run ID
        in: path
        name: runId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Approve a payroll run
      tags:
      - HR - Payroll
  /api/v1/hr/payroll/{runId}/calculate:
    post:
      parameters:
      - description: Payroll Run ID
        in: path
        name: runId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Calculate payroll for a run
      tags:
      - HR - Payroll
  /api/v1/hr/payroll/{runId}/payslip/{employeeId}:
    get:
      parameters:
      - description: Payroll Run ID
        in: path
        name: runId
        required: true
        type: string
      - description: Employee ID
        in: path
        name: employeeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a pay slip
      tags:
      - HR - Payroll
  /api/v1/hr/payroll/{runId}/process:
    post:
      parameters:
      - description: Payroll Run ID
        in: path
        name: runId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Process a payroll run
      tags:
      - HR - Payroll
...
securityDefinitions:
  BearerAuth:
    description: Enter your JWT token with the Bearer prefix, e.g. "Bearer eyJhbGciOiJI..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...