# gateway/handlers/hr.go
package handlers

import (
	"net/http"

	"github.com/gin-gonic/gin"
	"google.golang.org/grpc"
)

type HRHandler struct {
	conn *grpc.ClientConn
}

func NewHRHandler(conn *grpc.ClientConn) *HRHandler {
	return &HRHandler{conn: conn}
}

// CreateDepartment godoc
// @Summary      Create a new department
// @Tags         HR - Departments
// @Accept       json
// @Produce      json
// @Param        dept  body      map[string]interface{}  true  "Department data"
// @Success      201   {object}  map[string]interface{}
// @Failure      400   {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/departments [post]
func (h *HRHandler) CreateDepartment(c *gin.Context) {
	var req map[string]interface{}
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusCreated, gin.H{"message": "Department created", "data": req})
}

// GetDepartment godoc
// @Summary      Get a department by ID
// @Tags         HR - Departments
// @Produce      json
// @Param        id   path      string  true  "Department ID"
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/departments/{id} [get]
func (h *HRHandler) GetDepartment(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"id": c.Param("id")}) }

// ListDepartments godoc
// @Summary      List all departments
// @Tags         HR - Departments
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/departments [get]
func (h *HRHandler) ListDepartments(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"message": "List departments"}) }

// CreateEmployee godoc
// @Summary      Create a new employee
// @Tags         HR - Employees
// @Accept       json
// @Produce      json
// @Param        employee  body      map[string]interface{}  true  "Employee data"
// @Success      201       {object}  map[string]interface{}
// @Failure      400       {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/employees [post]
func (h *HRHandler) CreateEmployee(c *gin.Context) {
	var req map[string]interface{}
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusCreated, gin.H{"message": "Employee created", "data": req})
}

// GetEmployee godoc
// @Summary      Get an employee by ID
// @Tags         HR - Employees
// @Produce      json
// @Param        id   path      string  true  "Employee ID"
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/employees/{id} [get]
func (h *HRHandler) GetEmployee(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"id": c.Param("id")}) }

// ListEmployees godoc
// @Summary      List all employees
// @Tags         HR - Employees
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/employees [get]
func (h *HRHandler) ListEmployees(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"message": "List employees"}) }

// UpdateEmployee godoc
// @Summary      Update an employee
// @Tags         HR - Employees
// @Accept       json
// @Produce      json
// @Param        id        path      string                 true  "Employee ID"
// @Param        employee  body      map[string]interface{}  true  "Updated employee data"
// @Success      200       {object}  map[string]interface{}
// @Failure      400       {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/employees/{id} [put]
func (h *HRHandler) UpdateEmployee(c *gin.Context) {
	var req map[string]interface{}
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, gin.H{"message": "Employee updated", "id": c.Param("id"), "data": req})
}

// TerminateEmployee godoc
// @Summary      Terminate an employee
// @Tags         HR - Employees
// @Accept       json
// @Produce      json
// @Param        id   path      string                 true  "Employee ID"
// @Param        data body      map[string]interface{}  true  "Termination data"
// @Success      200  {object}  map[string]interface{}
// @Failure      400  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/employees/{id}/terminate [post]
func (h *HRHandler) TerminateEmployee(c *gin.Context) {
	var req map[string]interface{}
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, gin.H{"message": "Employee terminated", "id": c.Param("id")})
}

// SubmitLeaveRequest godoc
// @Summary      Submit a leave request
// @Tags         HR - Leave Management
// @Accept       json
// @Produce      json
// @Param        leave  body      map[string]interface{}  true  "Leave request data"
// @Success      201    {object}  map[string]interface{}
// @Failure      400    {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/leave/request [post]
func (h *HRHandler) SubmitLeaveRequest(c *gin.Context) {
	var req map[string]interface{}
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusCreated, gin.H{"message": "Leave request submitted", "data": req})
}

// ApproveLeave godoc
// @Summary      Approve a leave request
// @Tags         HR - Leave Management
// @Produce      json
// @Param        id   path      string  true  "Leave Request ID"
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/leave/{id}/approve [post]
func (h *HRHandler) ApproveLeave(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"message": "Leave approved", "id": c.Param("id")}) }

// RejectLeave godoc
// @Summary      Reject a leave request
// @Tags         HR - Leave Management
// @Produce      json
// @Param        id   path      string  true  "Leave Request ID"
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/leave/{id}/reject [post]
func (h *HRHandler) RejectLeave(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"message": "Leave rejected", "id": c.Param("id")}) }

// ListLeaveRequests godoc
// @Summary      List all leave requests
// @Tags         HR - Leave Management
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/leave/requests [get]
func (h *HRHandler) ListLeaveRequests(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"message": "List leave requests"}) }

// GetLeaveBalance godoc
// @Summary      Get leave balance for an employee
// @Tags         HR - Leave Management
// @Produce      json
// @Param        employeeId  path      string  true  "Employee ID"
// @Success      200         {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/leave/balance/{employeeId} [get]
func (h *HRHandler) GetLeaveBalance(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"employee_id": c.Param("employeeId")}) }

// ClockIn godoc
// @Summary      Clock in attendance
// @Tags         HR - Attendance
// @Accept       json
// @Produce      json
// @Param        data  body      map[string]interface{}  true  "Clock-in data"
// @Success      200   {object}  map[string]interface{}
// @Failure      400   {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/attendance/clock-in [post]
func (h *HRHandler) ClockIn(c *gin.Context) {
	var req map[string]interface{}
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, gin.H{"message": "Clocked in", "data": req})
}

// ClockOut godoc
// @Summary      Clock out attendance
// @Tags         HR - Attendance
// @Accept       json
// @Produce      json
// @Param        data  body      map[string]interface{}  true  "Clock-out data"
// @Success      200   {object}  map[string]interface{}
// @Failure      400   {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/attendance/clock-out [post]
func (h *HRHandler) ClockOut(c *gin.Context) {
	var req map[string]interface{}
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, gin.H{"message": "Clocked out", "data": req})
}

// ListAttendance godoc
// @Summary      List attendance records
// @Tags         HR - Attendance
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/attendance [get]
func (h *HRHandler) ListAttendance(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"message": "List attendance"}) }

// CreatePayrollRun godoc
// @Summary      Create a payroll run
// @Tags         HR - Payroll
// @Accept       json
// @Produce      json
// @Param        payroll  body      map[string]interface{}  true  "Payroll run data"
// @Success      201      {object}  map[string]interface{}
// @Failure      400      {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/payroll [post]
func (h *HRHandler) CreatePayrollRun(c *gin.Context) {
	var req map[string]interface{}
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusCreated, gin.H{"message": "Payroll run created", "data": req})
}

// CalculatePayroll godoc
// @Summary      Calculate payroll for a run
// @Tags         HR - Payroll
// @Produce      json
// @Param        runId  path      string  true  "Payroll Run ID"
// @Success      200    {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/payroll/{runId}/calculate [post]
func (h *HRHandler) CalculatePayroll(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"message": "Payroll calculated", "id": c.Param("runId")}) }

// ApprovePayroll godoc
// @Summary      Approve a payroll run
// @Tags         HR - Payroll
// @Produce      json
// @Param        runId  path      string  true  "Payroll Run ID"
// @Success      200    {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/payroll/{runId}/approve [post]
func (h *HRHandler) ApprovePayroll(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"message": "Payroll approved", "id": c.Param("runId")}) }

// ProcessPayroll godoc
// @Summary      Process a payroll run
// @Tags         HR - Payroll
// @Produce      json
// @Param        runId  path      string  true  "Payroll Run ID"
// @Success      200    {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/payroll/{runId}/process [post]
func (h *HRHandler) ProcessPayroll(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"message": "Payroll processed", "id": c.Param("runId")}) }

// ListPayrollRuns godoc
// @Summary      List all payroll runs
// @Tags         HR - Payroll
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/payroll [get]
func (h *HRHandler) ListPayrollRuns(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"message": "List payroll runs"}) }

// GetPayrollRun godoc
// @Summary      Get a payroll run by ID
// @Tags         HR - Payroll
// @Produce      json
// @Param        runId  path      string  true  "Payroll Run ID"
// @Success      200    {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/payroll/{runId} [get]
func (h *HRHandler) GetPayrollRun(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"id": c.Param("runId")}) }

// GetPaySlip godoc
// @Summary      Get a pay slip
// @Tags         HR - Payroll
// @Produce      json
// @Param        runId       path      string  true  "Payroll Run ID"
// @Param        employeeId  path      string  true  "Employee ID"
// @Success      200         {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/payroll/{runId}/payslip/{employeeId} [get]
func (h *HRHandler) GetPaySlip(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{"run_id": c.Param("runId"), "employee_id": c.Param("employeeId")})
}

// GetDashboard godoc
// @Summary      Get HR dashboard
// @Tags         HR - Dashboard
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Security     BearerAuth
// @Router       /api/v1/hr/dashboard [get]
func (h *HRHandler) GetDashboard(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"message": "HR dashboard"}) }